source/dev-requirements.txt
-r requirements.txt
pytest==7.1.3
pytest-django==4.5.2
pytest-cov==3.0.0
